// frontend/src/app/core/api/model/transactionCategoryRequest.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionCategoryRequestDto { 
    type: TransactionCategoryRequestDto.TypeEnum;
    name: string;
    systemKey?: string;
    active?: boolean;
    displayOrder: number;
}
export namespace TransactionCategoryRequestDto {
    export const TypeEnum = {
        Income: 'INCOME',
        Expense: 'EXPENSE'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}