// frontend/src/app/core/api/model/pageable.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PageableDto { 
    page?: number;
    size?: number;
    sort?: Array<string>;
}